Fixes for ENS Registry Contract

+++ contracts/ENS.sol
@@ -16,7 +16,7 @@ interface ENS {
   event NewResolver(bytes32 indexed node, address resolver);
 
   // Logged when the TTL of a node changes.
-  event NewTTL(bytes32 indexed node, uint64 ttl);
+  event NewTTL(bytes32 indexed node, uint256 ttl);
 
   function setSubnodeOwner(bytes32 node, bytes32 label, address owner) external;
 
@@ -24,11 +24,11 @@ interface ENS {
 
   function setOwner(bytes32 node, address owner) external;
 
-  function setTTL(bytes32 node, uint64 ttl) external;
+  function setTTL(bytes32 node, uint256 ttl) external;
 
   function getOwner(bytes32 node) external view returns (address);
 
   function getResolver(bytes32 node) external view returns (address);
 
-  function getTTL(bytes32 node) external view returns (uint64);
+  function getTTL(bytes32 node) external view returns (uint256);
 }
